# _15_appliance_store/docs/apis.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "store api docs",
    "version": "2.0"
  },
...
  "host": "localhost:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/employee": {
      "post": {
        "summary": "create employee",
        "operationId": "employee_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateEmployeeReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateEmployeeRequest"
            }
          }
        ],
        "tags": [
          "employee"
        ]
      }
    },
...
    "/api/v1/employee/{id}": {
      "get": {
        "summary": "get employee by id",
        "operationId": "employee_GetByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetEmployeeByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "uint64"
          }
        ],
        "tags": [
          "employee"
        ]
      },
      "delete": {
        "summary": "delete employee by id",
        "operationId": "employee_DeleteByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteEmployeeByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "uint64"
          }
        ],
        "tags": [
          "employee"
        ]
      },
      "put": {
        "summary": "update employee by id",
        "operationId": "employee_UpdateByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateEmployeeByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "uint64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "phone": {
                  "type": "string"
                },
                "role": {
                  "type": "integer",
                  "format": "int32"
                },
                "storeID": {
                  "type": "integer",
                  "format": "int64"
                },
                "passwordHash": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "employee"
        ]
      }
    },
    "/api/v1/employeePerformance": {
      "post": {
        "summary": "create employeePerformance",
        "operationId": "employeePerformance_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateEmployeePerformanceReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateEmployeePerformanceRequest"
            }
          }
        ],
        "tags": [
          "employeePerformance"
        ]
      }
    },
    "/api/v1/employeePerformance/calculate-commission": {
      "post": {
        "summary": "计算销售提成\n实现逻辑：\n1. 获取员工当月的销售记录\n2. 应用阶梯提成规则计算奖金\n3. 扣除已退货订单金额\n4. 生成提成明细记录",
        "operationId": "employeePerformance_CalculateSalesCommission",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CalculateSalesCommissionReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CalculateSalesCommissionRequest"
            }
          }
        ],
        "tags": [
          "employeePerformance"
        ]
      }
    },
    "/api/v1/employeePerformance/list": {
      "post": {
        "summary": "list of employeePerformance by query parameters",
        "operationId": "employeePerformance_List",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListEmployeePerformanceReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ListEmployeePerformanceRequest"
            }
          }
        ],
        "tags": [
          "employeePerformance"
        ]
      }
    },
    "/api/v1/employeePerformance/{id}": {
      "get": {
        "summary": "get employeePerformance by id",
        "operationId": "employeePerformance_GetByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetEmployeePerformanceByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "uint64"
          }
        ],
        "tags": [
          "employeePerformance"
        ]
      },
      "delete": {
        "summary": "delete employeePerformance by id",
        "operationId": "employeePerformance_DeleteByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteEmployeePerformanceByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "uint64"
          }
        ],
        "tags": [
          "employeePerformance"
        ]
      },
      "put": {
        "summary": "update employeePerformance by id",
        "operationId": "employeePerformance_UpdateByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateEmployeePerformanceByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "uint64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "employeeID": {
                  "type": "integer",
                  "format": "int64"
                },
                "month": {
                  "type": "string"
                },
                "salesAmount": {
                  "type": "string"
                },
                "commission": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "employeePerformance"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "typesColumn": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "exp": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "logic": {
          "type": "string"
        }
      }
    },
    "typesParams": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        },
        "sort": {
          "type": "string"
        },
        "columns": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/typesColumn"
          }
        }
      }
    },
...
    "v1CalculateSalesCommissionReply": {
      "type": "object",
      "properties": {
        "totalSales": {
          "type": "string"
        },
        "commissionAmount": {
          "type": "string"
        },
        "qualifiedOrders": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1CalculateSalesCommissionRequest": {
      "type": "object",
      "properties": {
        "employeeID": {
          "type": "integer",
          "format": "int64"
        },
        "month": {
          "type": "string"
        }
      }
    },
...
    "v1CreateEmployeePerformanceReply": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "uint64"
        }
      }
    },
    "v1CreateEmployeePerformanceRequest": {
      "type": "object",
      "properties": {
        "employeeID": {
          "type": "integer",
          "format": "int64"
        },
        "month": {
          "type": "string"
        },
        "salesAmount": {
          "type": "string"
        },
        "commission": {
          "type": "string"
        }
      }
    },
    "v1CreateEmployeeReply": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "uint64"
        }
      }
    },
    "v1CreateEmployeeRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "role": {
          "type": "integer",
          "format": "int32"
        },
        "storeID": {
          "type": "integer",
          "format": "int64"
        },
        "passwordHash": {
          "type": "string"
        }
      }
    },
...
    "v1DeleteEmployeeByIDReply": {
      "type": "object"
    },
    "v1DeleteEmployeePerformanceByIDReply": {
      "type": "object"
    },
...
    "v1Employee": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "uint64"
        },
        "name": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "role": {
          "type": "integer",
          "format": "int32"
        },
        "storeID": {
          "type": "integer",
          "format": "int64"
        },
        "passwordHash": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "v1EmployeePerformance": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "uint64"
        },
        "employeeID": {
          "type": "integer",
          "format": "int64"
        },
        "month": {
          "type": "string"
        },
        "salesAmount": {
          "type": "string"
        },
        "commission": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
...
    "v1GetEmployeeByIDReply": {
      "type": "object",
      "properties": {
        "employee": {
          "$ref": "#/definitions/v1Employee"
        }
      }
    },
    "v1GetEmployeePerformanceByIDReply": {
      "type": "object",
      "properties": {
        "employeePerformance": {
          "$ref": "#/definitions/v1EmployeePerformance"
        }
      }
    },
...
    "v1ListEmployeePerformanceReply": {
      "type": "object",
      "properties": {
        "total": {
          "type": "integer",
          "format": "int64"
        },
        "employeePerformances": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1EmployeePerformance"
          }
        }
      }
    },
    "v1ListEmployeePerformanceRequest": {
      "type": "object",
      "properties": {
        "params": {
          "$ref": "#/definitions/typesParams"
        }
      }
    },
...
    "v1UpdateEmployeeByIDReply": {
      "type": "object"
    },
    "v1UpdateEmployeePerformanceByIDReply": {
      "type": "object"
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "description": "Type Bearer your-jwt-token to Value",
      "name": "Authorization",
      "in": "header"
    }
  }
}